fix: faulty implicit cast of `float` variable to `long` (#13)
+ update gradle wrapper 6.8 -> 6.8.3

+++ server/src/main/java/Application.java
@@ -34,7 +34,7 @@ public class Application {
             } else {
                 // metric
             }
-            float startTime = System.currentTimeMillis();
+            long startTime = System.currentTimeMillis();
             world.process();
             time += System.currentTimeMillis() - startTime;
             if (runOnce) {
